move utils module, clean up comments
